requirements.txt
aniso8601==7.0.0
asgiref==3.5.0
Django==4.0.2
django-js-asset==2.0.0
django-mptt==0.13.4
djangorestframework==3.13.1
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
parso==0.8.3
Pillow==9.0.1
promise==2.3
psycopg2-binary==2.9.3
pytz==2021.3
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
whitenoise==5.3.0
